refactor(fullhistory/streaming): organize package -- doc map, layer-grouped filenames, split audit.go
No behavior change; pure organization. Kept as ONE package rather than
sub-packaging because the crash-injection hooks fire from INSIDE the real
catalog/protocol/sweep/ingest methods, so those must share a package to stay
package-private and keep the invariant tests meaningful.

- doc.go: new package architecture map (file -> layer), relocated from keys.go's
  package comment and expanded into a foundation -> catalog -> {config, freeze
  engine, ingestion} -> orchestration -> operability guide.
- Layer-grouped filenames (git mv, content unchanged):
    protocol.go -> catalog_protocol.go;  sweep.go -> catalog_sweep.go
    validate.go -> config_validate.go (+test);  lock.go -> config_lock.go (+test)
    build.go -> txindex.go (+test)
- Split the 853-line audit.go: types + Audit driver + RunAudit stay; the four
  invariant walks (INV-1..4) + filesystem helpers move to audit_invariants.go.
- gofmt: fix pre-existing unclean formatting in convergence_test.go /
  lifecycle_test.go / observability_test.go (whitespace only).
