+++ cmd/stellar-rpc/internal/fullhistory/streaming/catalog_protocol.go
