+++ cmd/stellar-rpc/internal/fullhistory/streaming/catalog_sweep.go
